Move all utilities to the utils folder Deleted unused utilities Move some utilities to the addons package
