+++ packages/react-data-grid-addons/src/data/RowSorter.js
@@ -1,6 +1,4 @@
-import { _utils } from 'react-data-grid';
-
-const { getMixedTypeValueRetriever, isImmutableCollection } = _utils;
+import { getMixedTypeValueRetriever, isImmutableCollection } from '../utils';
 
 export const comparer = (a, b) => {
   if (a > b) {
