+++ packages/react-data-grid/src/ColumnMetrics.ts
@@ -1,5 +1,5 @@
 export { sameColumn } from './ColumnComparer';
-import { isFrozen } from './ColumnUtils';
+import { isFrozen } from './utils/ColumnUtils';
 import { getScrollbarSize } from './utils';
 import { Column, CalculatedColumn, ColumnList, ColumnMetrics } from './common/types';
