+++ packages/react-data-grid/src/Row.tsx
@@ -3,8 +3,8 @@ import classNames from 'classnames';
 import shallowEqual from 'shallowequal';
 
 import Cell from './Cell';
-import { isFrozen } from './ColumnUtils';
-import * as rowUtils from './RowUtils';
+import { isFrozen } from './utils/ColumnUtils';
+import * as rowUtils from './utils/rowUtils';
 import { RowRenderer, RowRendererProps, CellRenderer, CellRendererProps, CalculatedColumn } from './common/types';
 
 export default class Row<R> extends React.Component<RowRendererProps<R>> implements RowRenderer<R> {
