+++ packages/react-data-grid/src/index.ts
@@ -1,9 +1,8 @@
 export { default, ReactDataGridProps } from './ReactDataGrid';
-export { default as rowComparer } from './common/utils/RowComparer';
+export { default as rowComparer } from './utils/RowComparer';
 export { default as Cell } from './Cell';
 export { default as Row } from './Row';
 export { default as HeaderCell } from './HeaderCell';
-import * as _utils from './common/utils';
 import * as _helpers from './helpers';
 export * from './Columns';
 export * from './formatters';
@@ -11,6 +10,5 @@ export * from './common/editors';
 export * from './common/enums';
 export * from './common/types';
 export {
-  _utils,
   _helpers
 };
