+++ packages/react-data-grid/src/utils/canvasUtils.ts
@@ -1,4 +1,4 @@
-import { isFrozen } from '../ColumnUtils';
+import { isFrozen } from './ColumnUtils';
 import { CalculatedColumn } from '../common/types';
 
 export function getColumnScrollPosition<R>(columns: CalculatedColumn<R>[], idx: number, currentScrollLeft: number, currentClientWidth: number): number {
